requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==2.0.12
Django==3.2.9
djangorestframework==3.12.4
flake8==7.1.0
gunicorn==20.1.0
idna==3.7
mccabe==0.7.0
pycodestyle==2.12.0
pyflakes==3.2.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
setuptools==71.0.1
sqlparse==0.5.1
urllib3==1.26.19
gunicorn==20.1.0
psycopg2-binary==2.9.3
dj-database-url==0.5.0